requirements.txt
fastapi
python-dotenv
uvicorn[standard]
python-multipart